refactor: [sync general] reorganize components related to field section
