+++ src/components/analytics/VisualizationType.js
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import { RadioGroup } from './RadioGroup'
-import { FormSection } from '../sections/general/form-sections'
+import { FieldSection } from '../field'
 import {
     createInitialValues,
     updateAttributesList,
@@ -84,7 +84,7 @@ export const VisualizationType = ({
     }
 
     return (
-        <FormSection>
+        <FieldSection>
             <RadioGroup
                 required
                 name="type"
@@ -94,7 +94,7 @@ export const VisualizationType = ({
                 options={visualizationOptions}
                 defaultValues={defaultValues}
             />
-        </FormSection>
+        </FieldSection>
     )
 }
