+++ src/components/field/GlobalProgramCompletion.js
@@ -2,8 +2,8 @@ import React from 'react'
 import cx from 'classnames'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
-import { CheckboxField } from '../CheckboxField'
-import styles from '../../styles/Form.module.css'
+import { CheckboxField } from './CheckboxField'
+import styles from './Fied.module.css'
 
 export const GlobalProgramCompletion = ({ disable, settings, onChange }) => {
     const handleChange = e => {
