+++ src/components/field/TextField.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from '@dhis2/prop-types'
 import { InputField } from '@dhis2/ui'
-import { FormSection } from '../sections/general/form-sections'
+import { FieldSection } from './FieldSection'
 
 export const TextField = ({ label, helpText, value, ...props }) => (
-    <FormSection>
+    <FieldSection>
         <InputField
             inputWidth="300px"
             helpText={helpText}
@@ -12,7 +12,7 @@ export const TextField = ({ label, helpText, value, ...props }) => (
             value={value}
             {...props}
         />
-    </FormSection>
+    </FieldSection>
 )
 
 TextField.propTypes = {
