+++ src/components/field/index.js
@@ -3,6 +3,7 @@ export * from './ButtonField'
 export * from './CheckboxField'
 export * from './DataSync'
 export * from './EncryptDB'
+export * from './FieldSection'
 export * from './GlobalProgramCompletion'
 export * from './MatomoId'
 export * from './MatomoUrl'
