--- src/components/sections/general/form-sections/form-section.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import PropTypes from '@dhis2/prop-types'
-import formStyles from '../../../../styles/Form.module.css'
-
-export const FormSection = props => (
-    <div className={formStyles.row}>{props.children}</div>
-)
-
-FormSection.propTypes = {
-    children: PropTypes.element,
-}
