--- src/components/sections/general/form-sections/index.js
@@ -1,6 +1,0 @@
-export * from './form-section'
-export * from './button-field'
-export * from './sync-checkbox-field'
-export * from './sync-input-number-field'
-export * from './sync-input-phone-field'
-export * from './sync-select-field'
